requirements.txt
pandas==2.1.3
torch==2.4.1
langchain==0.3.2
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-groq==0.2.0
langgraph==0.2.56
beautifulsoup4==4.12.3
lxml==4.9.3
python-dotenv==1.0.0
readability-lxml==0.8.1
tqdm==4.66.1